Refactor to replace lrtd with detectors
