+++ src/detectors.rs
